1. Add properties to the application.properties and use them in ApplicationSecurityConfig 2. Change qualifier to postgesql for db identification 3. Add constructor injection 4. Change a logic slightly in CartController and ProductController

+++ src/main/java/com/obolonyk/shopboot/dao/ProductDao.java
@@ -8,7 +8,7 @@ public interface ProductDao {
 
     List<Product> getAll();
 
-    Product getById(int id);
+    Product getById(Long id);
 
     void save(Product product);
